requirements.txt
numpy>=1.16.0
opencv-python>=4.1.0
matplotlib>=3.1.0
scipy>=1.3.0
pyyaml>=5.1.0

# Optional, for PCD file support
# open3d>=0.9.0